initial attempt to split up
