+++ esc.hpp
@@ -1,34 +1,6 @@
 #pragma once
 
-#include <pigpio.h>
-#include <iostream>
-#include <csignal>
-#include <cstdlib>
-#include <functional>
-
-class Servo {
-	public:
-		Servo(int pin);
-		int getPulseWidth();
-		void setPulseWidth(int pulseWidth);
-
-	protected:
-		int __pin;
-};
-
-class AngularServo : public Servo {
-	public:
-		AngularServo(int pin, float minAngle, float maxAngle, int minPulseWidthUs, int maxPulseWidthUs);
-		void setAngle(float angle);
-		int getAngle();
-
-	protected:
-		float __angle;
-		float __minAngle;
-		float __maxAngle;
-		int __minPulseWidthUs;
-		int __maxPulseWidthUs;
-};
+#include "angular_servo.hpp"
 
 class ESC : public AngularServo {
 	public:
